requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.0
